refactor: change structure Attendance / yoonho
