+++ Split/Storyboard/Split/Storyboard/Base.lproj/Attendance.storyboard
@@ -1,0 +1,84 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<document type="com.apple.InterfaceBuilder3.CocoaTouch.Storyboard.XIB" version="3.0" toolsVersion="17156" targetRuntime="iOS.CocoaTouch" propertyAccessControl="none" useAutolayout="YES" useTraitCollections="YES" useSafeAreas="YES" colorMatched="YES" initialViewController="jb5-Zs-uuO">
+    <device id="retina6_1" orientation="portrait" appearance="light"/>
+    <dependencies>
+        <plugIn identifier="com.apple.InterfaceBuilder.IBCocoaTouchPlugin" version="17126"/>
+        <capability name="Safe area layout guides" minToolsVersion="9.0"/>
+        <capability name="System colors in document resources" minToolsVersion="11.0"/>
+        <capability name="documents saved in the Xcode 8 format" minToolsVersion="8.0"/>
+    </dependencies>
+    <scenes>
+        <!--Attendance View Controller-->
+        <scene sceneID="JLO-xe-VNW">
+            <objects>
+                <viewController storyboardIdentifier="attendanceView" id="jb5-Zs-uuO" customClass="AttendanceViewController" customModule="Split" customModuleProvider="target" sceneMemberID="viewController">
+                    <view key="view" contentMode="scaleToFill" id="fgw-qT-zox">
+                        <rect key="frame" x="0.0" y="0.0" width="414" height="896"/>
+                        <autoresizingMask key="autoresizingMask" widthSizable="YES" heightSizable="YES"/>
+                        <subviews>
+                            <button opaque="NO" contentMode="scaleToFill" contentHorizontalAlignment="center" contentVerticalAlignment="center" buttonType="system" lineBreakMode="middleTruncation" translatesAutoresizingMaskIntoConstraints="NO" id="icK-8h-KLN">
+                                <rect key="frame" x="192" y="433" width="30" height="30"/>
+                                <state key="normal" title="QR"/>
+                                <connections>
+                                    <action selector="showScanner" destination="jb5-Zs-uuO" eventType="touchUpInside" id="BDH-1x-X10"/>
+                                </connections>
+                            </button>
+                        </subviews>
+                        <viewLayoutGuide key="safeArea" id="QOY-GD-rJo"/>
+                        <color key="backgroundColor" systemColor="systemBackgroundColor"/>
+                        <constraints>
+                            <constraint firstItem="icK-8h-KLN" firstAttribute="centerY" secondItem="fgw-qT-zox" secondAttribute="centerY" id="D7E-9v-S1K"/>
+                            <constraint firstItem="icK-8h-KLN" firstAttribute="centerX" secondItem="fgw-qT-zox" secondAttribute="centerX" id="ypl-x1-F5D"/>
+                        </constraints>
+                    </view>
+                    <navigationItem key="navigationItem" id="cWO-qV-7UO"/>
+                </viewController>
+                <placeholder placeholderIdentifier="IBFirstResponder" id="cLG-zJ-3hc" userLabel="First Responder" customClass="UIResponder" sceneMemberID="firstResponder"/>
+            </objects>
+            <point key="canvasLocation" x="3117.3913043478265" y="-2096.6517857142858"/>
+        </scene>
+        <!--Attendance Completion View Controller-->
+        <scene sceneID="ICc-dE-ARh">
+            <objects>
+                <viewController storyboardIdentifier="attendanceViewController" id="K0x-m7-Ymt" customClass="AttendanceCompletionViewController" customModule="Split" customModuleProvider="target" sceneMemberID="viewController">
+                    <view key="view" contentMode="scaleToFill" id="CeS-gm-kSA">
+                        <rect key="frame" x="0.0" y="0.0" width="414" height="896"/>
+                        <autoresizingMask key="autoresizingMask" widthSizable="YES" heightSizable="YES"/>
+                        <subviews>
+                            <label opaque="NO" userInteractionEnabled="NO" contentMode="left" horizontalHuggingPriority="251" verticalHuggingPriority="251" text="complete" textAlignment="center" lineBreakMode="tailTruncation" numberOfLines="0" baselineAdjustment="alignBaselines" adjustsFontSizeToFit="NO" translatesAutoresizingMaskIntoConstraints="NO" id="5VD-W8-F5P">
+                                <rect key="frame" x="20" y="438" width="374" height="20.5"/>
+                                <fontDescription key="fontDescription" type="system" pointSize="17"/>
+                                <nil key="textColor"/>
+                                <nil key="highlightedColor"/>
+                            </label>
+                            <button opaque="NO" contentMode="scaleToFill" contentHorizontalAlignment="center" contentVerticalAlignment="center" buttonType="system" lineBreakMode="middleTruncation" translatesAutoresizingMaskIntoConstraints="NO" id="zuz-FK-4IA">
+                                <rect key="frame" x="173.5" y="732" width="67" height="30"/>
+                                <state key="normal" title="Complete"/>
+                            </button>
+                        </subviews>
+                        <viewLayoutGuide key="safeArea" id="LtA-UV-qfm"/>
+                        <color key="backgroundColor" systemColor="systemBackgroundColor"/>
+                        <constraints>
+                            <constraint firstItem="zuz-FK-4IA" firstAttribute="centerX" secondItem="CeS-gm-kSA" secondAttribute="centerX" id="50L-r8-lLw"/>
+                            <constraint firstItem="LtA-UV-qfm" firstAttribute="bottom" secondItem="zuz-FK-4IA" secondAttribute="bottom" constant="100" id="9sw-yz-DR6"/>
+                            <constraint firstItem="5VD-W8-F5P" firstAttribute="leading" secondItem="LtA-UV-qfm" secondAttribute="leading" constant="20" id="JRX-ff-5cR"/>
+                            <constraint firstItem="LtA-UV-qfm" firstAttribute="trailing" secondItem="5VD-W8-F5P" secondAttribute="trailing" constant="20" id="hPW-4N-XKh"/>
+                            <constraint firstItem="5VD-W8-F5P" firstAttribute="centerY" secondItem="CeS-gm-kSA" secondAttribute="centerY" id="iNX-7L-xuZ"/>
+                        </constraints>
+                    </view>
+                    <connections>
+                        <outlet property="completionButton" destination="zuz-FK-4IA" id="Ipk-du-VE5"/>
+                        <outlet property="urlLabel" destination="5VD-W8-F5P" id="Ei5-Ks-t2v"/>
+                    </connections>
+                </viewController>
+                <placeholder placeholderIdentifier="IBFirstResponder" id="0yK-Uq-zY8" userLabel="First Responder" customClass="UIResponder" sceneMemberID="firstResponder"/>
+            </objects>
+            <point key="canvasLocation" x="4005.7971014492755" y="-2097.3214285714284"/>
+        </scene>
+    </scenes>
+    <resources>
+        <systemColor name="systemBackgroundColor">
+            <color white="1" alpha="1" colorSpace="custom" customColorSpace="genericGamma22GrayColorSpace"/>
+        </systemColor>
+    </resources>
+</document>
